Moved debug rendering of TerrainPager tiles to tool
